1.增加响应 下载模板请求时，追加回复多个图片功能 2.重构了对模板目录文件信息的缓存功能 3.规范了代码格式，删除无用代码和注释

+++ virtualscannerserverlib/chat_client.cpp
@@ -148,7 +148,7 @@ namespace DROWNINGLIU
 					char *tmp = NULL;
 					char accout[11] = { 0 };
 					char passWd[20] = "qgdyuwgqu455";
-					uint8_t passWdSize = strlen(passWd);
+					uint8_t passWdSize = (uint8_t)strlen(passWd);
 					reqPackHead_t head;
 					char *tmpSendPackDataOne = szBuf, *sendPackData = szBuf2;
 					int  checkCode = 0, contenLenth = 0;
